[rollout] refactor: move server management out of AgentLoopManager
