+++ tests/checkpoint_engine/test_special_server_adapter.py
@@ -21,13 +21,12 @@ from transformers import PreTrainedTokenizer
 
 from tests.checkpoint_engine.test_utils import create_trainer_worker_group
 from verl.checkpoint_engine import CheckpointEngineManager
-from verl.experimental.agent_loop.agent_loop import AgentLoopManager, AsyncLLMServerManager
-from verl.experimental.fully_async_policy.agent_loop.agent_loop import FullyAsyncLLMServerManager
 from verl.single_controller.ray import (
     RayResourcePool,
 )
 from verl.utils.config import omega_conf_to_dataclass
 from verl.workers.config import CheckpointEngineConfig, HFModelConfig
+from verl.workers.rollout.llm_server import FullyLLMServerClient, LLMServerClient, LLMServerManager
 
 
 @pytest.fixture
@@ -55,7 +54,7 @@ def init_config() -> DictConfig:
 
 
 async def _run_update_weights_with_global_steps_none(
-    server_manager: AsyncLLMServerManager,
+    server_manager: LLMServerClient,
     checkpoint_manager: CheckpointEngineManager,
     tokenizer: PreTrainedTokenizer,
 ):
@@ -83,7 +82,7 @@ async def _run_update_weights_with_global_steps_none(
 async def _run_server_manager_without_resume(
     initial_steps: int,
     train_steps: int,
-    server_manager: AsyncLLMServerManager,
+    server_manager: LLMServerClient,
     checkpoint_manager: CheckpointEngineManager,
     prompts: list[list[dict]],
     tokenizer: PreTrainedTokenizer,
@@ -124,7 +123,7 @@ async def _run_server_manager_without_resume(
 async def _run_server_manager_with_resume(
     initial_steps: int,
     train_steps: int,
-    server_manager: FullyAsyncLLMServerManager,
+    server_manager: FullyLLMServerClient,
     checkpoint_manager: CheckpointEngineManager,
     prompts: list[list[dict]],
     tokenizer: PreTrainedTokenizer,
@@ -195,19 +194,11 @@ async def test_server_adapter(init_config):
     trainer.reset()
 
     # 2. create standalone rollout with AgentLoopManager
-    agent_loop_manager = await AgentLoopManager.create(config=init_config)
-    servers = list(
-        zip(
-            agent_loop_manager.server_addresses,
-            [server._server_handle for server in agent_loop_manager.rollout_replicas],
-            strict=True,
-        )
-    )
-    load_balancer_handle = agent_loop_manager.global_load_balancer
+    llm_server_manager = await LLMServerManager.create(config=init_config)
 
     # 3. create checkpoint engine manager
     checkpoint_manager = CheckpointEngineManager(
-        config=checkpoint_engine_config, trainer=trainer, replicas=agent_loop_manager.rollout_replicas
+        config=checkpoint_engine_config, trainer=trainer, replicas=llm_server_manager.get_replicas()
     )
 
     n = 4
@@ -218,36 +209,29 @@ async def test_server_adapter(init_config):
         [{"role": "user", "content": "Please write an article about the geography of America, at least 1000 words."}],
     ] * n
 
-    server_manager = AsyncLLMServerManager(
-        config=init_config, servers=servers, load_balancer_handle=load_balancer_handle
-    )
-
     # 4. test update_weights with global_steps=None
     await _run_update_weights_with_global_steps_none(
-        server_manager=server_manager,
+        server_manager=llm_server_manager.get_client(),
         checkpoint_manager=checkpoint_manager,
         tokenizer=model_config.tokenizer,
     )
 
-    # 5. test AsyncLLMServerManager without partial rollout resume
+    # 5. test LLMServerClient without partial rollout resume
     await checkpoint_manager.update_weights(global_steps=0)
     await _run_server_manager_without_resume(
         initial_steps=1,
         train_steps=3,
-        server_manager=server_manager,
+        server_manager=llm_server_manager.get_client(),
         checkpoint_manager=checkpoint_manager,
         prompts=prompts,
         tokenizer=model_config.tokenizer,
     )
 
-    # 6. test FullyAsyncLLMServerManager with partial rollout resume
-    server_manager = FullyAsyncLLMServerManager(
-        config=init_config, servers=servers, load_balancer_handle=load_balancer_handle
-    )
+    # 6. test FullyLLMServerClient with partial rollout resume
     await _run_server_manager_with_resume(
         initial_steps=4,
         train_steps=3,
-        server_manager=server_manager,
+        server_manager=llm_server_manager.get_client(fully_async=True),
         checkpoint_manager=checkpoint_manager,
         prompts=prompts,
         tokenizer=model_config.tokenizer,
