+++ verl/experimental/agent_loop/__init__.py
@@ -17,7 +17,6 @@ from .agent_loop import (
     AgentLoopManager,
     AgentLoopOutput,
     AgentLoopWorker,
-    AsyncLLMServerManager,
     get_trajectory_info,
 )
 from .single_turn_agent_loop import SingleTurnAgentLoop
@@ -28,7 +27,6 @@ _ = [SingleTurnAgentLoop, ToolAgentLoop]
 __all__ = [
     "AgentLoopBase",
     "AgentLoopManager",
-    "AsyncLLMServerManager",
     "AgentLoopWorker",
     "AgentLoopOutput",
     "get_trajectory_info",
