+++ verl/experimental/fully_async_policy/config/fully_async_ppo_megatron_trainer.yaml
@@ -19,7 +19,7 @@ async_training:
   require_batches: 1
 
   # When synchronizing parameters, Whether to resume generation when rollout is interrupted.
-  # If True, AsyncLLMServerManager auto resume generation, making rollout interruption invisible to the AgentLoop.
+  # If True, LLMServerClient auto resume generation, making rollout interruption invisible to the AgentLoop.
   partial_rollout: True
 
   # whether to use trainer do_validate
