+++ verl/workers/rollout/llm_server.py
@@ -1,0 +1,381 @@
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+Utility classes for manage and request LLM servers:
+- LLMServerManager: manage life-cycle of LLM servers, including launch, tear-down replicas.
+- LLMServerClient: proxy client to request LLM servers, used by AgentLoopWorker.
+- GlobalRequestLoadBalancer: global load balancer for LLMServerClient.
+"""
+
+import asyncio
+import logging
+import os
+from typing import Any, Optional
+from uuid import uuid4
+
+import ray
+import torch
+from cachetools import LRUCache
+from omegaconf import DictConfig
+
+from verl.single_controller.ray.base import RayResourcePool, RayWorkerGroup
+from verl.utils.ray_utils import auto_await
+from verl.utils.rollout_trace import rollout_trace_op
+from verl.utils.tokenizer import normalize_token_ids
+from verl.workers.rollout.replica import DiffusionOutput, RolloutReplica, TokenOutput, get_rollout_replica_class
+from verl.workers.rollout.utils import update_prometheus_config
+
+logger = logging.getLogger(__file__)
+logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
+
+DEFAULT_ROUTING_CACHE_SIZE = 10000
+
+
+@ray.remote
+class GlobalRequestLoadBalancer:
+    """Global sticky-session + in-flight load balancer shared by all AgentLoopWorkers."""
+
+    def __init__(self, servers: dict[str, ray.actor.ActorHandle], max_cache_size: int = DEFAULT_ROUTING_CACHE_SIZE):
+        if not servers:
+            raise ValueError("server must be non-empty")
+
+        self._server = servers
+        self._inflight_requests: dict[str, int] = {sid: 0 for sid in servers}
+        self._request_id_to_server: LRUCache = LRUCache(maxsize=max_cache_size)
+
+    def acquire_server(self, request_id: str) -> str:
+        """Acquire a server for the given request, reusing the same server for multi-turn conversations."""
+        # request-level sticky (multi-turn: same conversation -> same server)
+        if request_id in self._request_id_to_server:
+            server_id = self._request_id_to_server[request_id]
+            self._inflight_requests[server_id] += 1
+            return server_id
+
+        # new request: route to least loaded server
+        server_id = min(self._inflight_requests, key=self._inflight_requests.get)
+        self._request_id_to_server[request_id] = server_id
+        self._inflight_requests[server_id] += 1
+        return server_id
+
+    def release_server(self, server_id: str) -> None:
+        """Release a server after a request completes, decrementing its inflight count."""
+        if server_id not in self._inflight_requests:
+            raise ValueError(f"Invalid server_id for release: {server_id}")
+        if self._inflight_requests[server_id] <= 0:
+            raise ValueError(f"Release called with no inflight requests on server {server_id}")
+        self._inflight_requests[server_id] -= 1
+
+    def add_servers(self, servers: dict[str, ray.actor.ActorHandle]) -> None:
+        """Add new servers to the server handles."""
+        raise NotImplementedError("Not implemented")
+
+    def remove_servers(self, server_ids: list[str]) -> None:
+        """Remove servers from the server handles."""
+        raise NotImplementedError("Not implemented")
+
+
+class LLMServerClient:
+    """
+    A class to manage multiple OpenAI compatible LLM servers. This class provides
+    - Load balance: least in-flight requests load balancing via global coordination
+    - Sticky session: send multi-turn chat completions to same server for automatic prefix caching
+    """
+
+    def __init__(
+        self,
+        config: DictConfig,
+        servers: dict[str, ray.actor.ActorHandle],
+        load_balancer_handle: ray.actor.ActorHandle,
+    ):
+        """Initialize the LLMServerClient.
+
+        Args:
+            config (DictConfig): whole config for main entrypoint.
+            servers (dict[str, ray.actor.ActorHandle]): handle for each LLM server.
+            load_balancer_handle (ray.actor.ActorHandle): shared global load balancer actor.
+        """
+        self.config = config
+        self._load_balancer = load_balancer_handle
+        self._server_id_to_handle: dict[str, ray.actor.ActorHandle] = servers
+
+    async def _acquire_server(self, request_id: str) -> tuple[str, ray.actor.ActorHandle]:
+        server_id = await self._load_balancer.acquire_server.remote(request_id=request_id)
+        handle = self._server_id_to_handle.get(server_id)
+        if handle is None:
+            raise RuntimeError(f"Unknown server_id returned by load balancer: {server_id}")
+        return server_id, handle
+
+    def _release_server(self, server_id: str) -> None:
+        # Fire-and-forget: release is just a counter decrement, no need to await.
+        # Awaiting here risks blocking the finally clause if the LB actor is unresponsive.
+        self._load_balancer.release_server.remote(server_id=server_id)
+
+    @rollout_trace_op
+    async def generate(
+        self,
+        request_id,
+        *,
+        prompt_ids: list[int],
+        sampling_params: dict[str, Any],
+        image_data: Optional[list[Any]] = None,
+        video_data: Optional[list[Any]] = None,
+        **kwargs: Any,
+    ) -> TokenOutput | DiffusionOutput:
+        """Generate tokens from prompt ids.
+
+        Args:
+            request_id (str): request id for sticky session.
+            prompt_ids (List[int]): List of prompt token ids.
+            sampling_params (Dict[str, Any]): Sampling parameters for the chat completion.
+
+        Returns:
+            TokenOutput | DiffusionOutput: token or diffusion output
+        """
+        server_id, server = await self._acquire_server(request_id)
+        try:
+            output: TokenOutput | DiffusionOutput = await server.generate.remote(
+                request_id=uuid4().hex,  # use new request_id for each turn
+                prompt_ids=prompt_ids,
+                sampling_params=sampling_params,
+                image_data=image_data,
+                video_data=video_data,
+                **kwargs,
+            )
+            return output
+        finally:
+            self._release_server(server_id)
+
+
+class FullyLLMServerClient(LLMServerClient):
+    """FullyLLMServerClient supports resume generation on partial rollout, making rollout interruption
+    invisible to the AgentLoop.
+    """
+
+    @rollout_trace_op
+    async def generate(
+        self,
+        request_id,
+        *,
+        prompt_ids: list[int],
+        sampling_params: dict[str, Any],
+        image_data: Optional[list[Any]] = None,
+        video_data: Optional[list[Any]] = None,
+    ) -> TokenOutput:
+        """Generate tokens from prompt ids.
+
+        Args:
+            request_id (str): request id for sticky session.
+            prompt_ids (List[int]): List of prompt token ids.
+            sampling_params (Dict[str, Any]): Sampling parameters for the chat completion.
+            image_data (Optional[List[Any]]): Image data for the chat completion.
+            video_data (Optional[List[Any]]): Video data for the chat completion.
+
+        Returns:
+            TokenOutput: token output
+        """
+        prompt_ids = normalize_token_ids(prompt_ids)
+
+        limit_key = None
+        if "max_tokens" in sampling_params:
+            limit_key = "max_tokens"
+        elif "max_new_tokens" in sampling_params:
+            limit_key = "max_new_tokens"
+        original_max_tokens = sampling_params.get(limit_key) if limit_key else None
+
+        final_output = TokenOutput(
+            token_ids=[],
+            log_probs=[],
+            num_preempted=0,
+        )
+        min_global_steps, max_global_steps = None, None
+
+        while True:
+            # 1. generate tokens
+            output = await super().generate(
+                request_id=request_id,
+                prompt_ids=prompt_ids + final_output.token_ids,
+                sampling_params=sampling_params,
+                image_data=image_data,
+                video_data=video_data,
+            )
+
+            # 2. merge output into final_output
+            final_output.token_ids.extend(output.token_ids)
+            if output.log_probs is not None:
+                final_output.log_probs.extend(output.log_probs)
+            # On partial rollout resume the model version may differ, so keep
+            # existing routing and only append routing for newly generated tokens.
+            if output.routed_experts is not None and len(output.token_ids) > 0:
+                if final_output.routed_experts is None:
+                    final_output.routed_experts = output.routed_experts
+                else:
+                    final_output.routed_experts = torch.cat(
+                        [final_output.routed_experts, output.routed_experts[-len(output.token_ids) :]],
+                        dim=0,
+                    )
+            if output.num_preempted is not None:
+                final_output.num_preempted += output.num_preempted
+            final_output.stop_reason = output.stop_reason
+
+            # update model weights version
+            global_steps = output.extra_fields.get("global_steps", None)
+            if min_global_steps is None:
+                min_global_steps = global_steps
+            max_global_steps = global_steps
+
+            # 3. update max_new_tokens
+            if original_max_tokens is not None:
+                sampling_params[limit_key] = original_max_tokens - len(final_output.token_ids)
+                if len(final_output.token_ids) >= original_max_tokens:
+                    final_output.stop_reason = "length"
+                    break
+
+            # 4. check stop reason
+            if output.stop_reason not in ("aborted", "abort") or not self.config.async_training.partial_rollout:
+                break
+        final_output.extra_fields["global_steps"] = global_steps
+        final_output.extra_fields["min_global_steps"] = min_global_steps
+        final_output.extra_fields["max_global_steps"] = max_global_steps
+        return final_output
+
+
+class LLMServerManager:
+    """LLMServerManager is responsible for:
+    - Launch server replicas
+    - Launch global load balancer
+    - Elastic launch/tear-down new replicas
+
+    Args:
+        config (DictConfig): Config for the trainer entrypoint.
+        worker_group (RayWorkerGroup): Worker group for the server replicas. If not none, init hybrid server,
+            else init standalone server with a new resource pool.
+        rollout_resource_pool (RayResourcePool): Resource pool for the server replicas, only needed for TensorRT-LLM.
+    """
+
+    def __init__(
+        self,
+        config: DictConfig,
+        worker_group: RayWorkerGroup = None,
+        rollout_resource_pool: RayResourcePool = None,
+    ):
+        self.config = config
+        self.rollout_config = config.actor_rollout_ref.rollout
+        self.model_config = config.actor_rollout_ref.model
+        self.worker_group = worker_group
+        self.rollout_resource_pool = rollout_resource_pool
+
+        assert worker_group is not None or self.rollout_config.nnodes > 0, "nnodes must be > 0 in standalone mode"
+
+        # for recipe to change
+        if not hasattr(self, "rollout_replica_class"):
+            self.rollout_replica_class = get_rollout_replica_class(self.rollout_config.name)
+
+    @classmethod
+    @auto_await
+    async def create(cls, *args, **kwargs):
+        """Create the LLMServerManager."""
+        instance = cls(*args, **kwargs)
+        await instance._initialize_llm_servers()
+        await instance._init_global_load_balancer()
+        return instance
+
+    async def _initialize_llm_servers(self):
+        """Initialize the LLM server replicas."""
+        rollout_world_size = (
+            self.rollout_config.tensor_model_parallel_size
+            * self.rollout_config.data_parallel_size
+            * self.rollout_config.pipeline_model_parallel_size
+        )
+        world_size = (
+            self.worker_group.world_size
+            if self.worker_group
+            else self.rollout_config.n_gpus_per_node * self.rollout_config.nnodes
+        )
+        num_replicas = world_size // rollout_world_size
+
+        self.rollout_replicas = [
+            self.rollout_replica_class(
+                replica_rank=replica_rank,
+                config=self.rollout_config,
+                model_config=self.model_config,
+                gpus_per_node=self.rollout_config.n_gpus_per_node,
+            )
+            for replica_rank in range(num_replicas)
+        ]
+
+        if self.worker_group and self.rollout_config.name != "trtllm":
+            await asyncio.gather(*[server.init_hybrid(self.worker_group) for server in self.rollout_replicas])
+        # TODO: unify trtllm to init_hybrid
+        elif self.worker_group and self.rollout_config.name == "trtllm":
+            await asyncio.gather(
+                *[
+                    server.init_hybrid_colocated(self.worker_group, self.rollout_resource_pool)
+                    for server in self.rollout_replicas
+                ]
+            )
+        else:
+            await asyncio.gather(*[server.init_standalone() for server in self.rollout_replicas])
+
+        self.server_handles = [server._server_handle for server in self.rollout_replicas]
+        self.server_addresses = [server._server_address for server in self.rollout_replicas]
+        print(f"LLMServerManager: {self.server_addresses}")
+
+        # Update Prometheus configuration with server addresses
+        if self.rollout_config.prometheus.enable:
+            if self.rollout_config.disable_log_stats:
+                raise ValueError("PROMETHEUS needs disable_log_stats==False, but it is currently True.")
+            update_prometheus_config(self.rollout_config.prometheus, self.server_addresses, self.rollout_config.name)
+
+    async def _init_global_load_balancer(self) -> None:
+        self.global_load_balancer = GlobalRequestLoadBalancer.remote(
+            servers=dict(zip(self.server_addresses, self.server_handles, strict=True)),
+            max_cache_size=DEFAULT_ROUTING_CACHE_SIZE,
+        )
+
+    def get_client(self, fully_async: bool = False) -> LLMServerClient:
+        """Get the LLMServerClient to request LLM server replicas.
+
+        Args:
+            fully_async (bool): Whether to return the FullyLLMServerClient.
+        """
+        servers = dict(zip(self.server_addresses, self.server_handles, strict=True))
+        if not fully_async:
+            return LLMServerClient(config=self.config, servers=servers, load_balancer_handle=self.global_load_balancer)
+        else:
+            return FullyLLMServerClient(
+                config=self.config, servers=servers, load_balancer_handle=self.global_load_balancer
+            )
+
+    def get_addresses(self) -> list[str]:
+        """Get the OpenAI chat completion API http addresses of the LLM server replicas."""
+        return self.server_addresses
+
+    def get_replicas(self) -> list[RolloutReplica]:
+        """Get the LLM server replicas."""
+        return self.rollout_replicas
+
+    @auto_await
+    async def clear_kv_cache(self):
+        """Clear all rollout kv cache, but don`t sleep."""
+        await asyncio.gather(*[replica.clear_kv_cache() for replica in self.rollout_replicas])
+
+    @auto_await
+    async def start_profile(self, **kwargs):
+        """Start profiling on all rollout replicas."""
+        await asyncio.gather(*[replica.start_profile(**kwargs) for replica in self.rollout_replicas])
+
+    @auto_await
+    async def stop_profile(self):
+        """Stop profiling on all rollout replicas."""
+        await asyncio.gather(*[replica.stop_profile() for replica in self.rollout_replicas])
